ml_python/requirements.txt
absl-py==1.0.0
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
datasets==1.0.2
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
filelock==3.4.1
fire==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.4.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
jellyfish==0.9.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
lime==0.2.0.1
llvmlite==0.36.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
nltk==3.6.7
notebook==6.4.8
numba==0.53.1
numpy==1.19.2
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.6.2
Pygments==2.11.2
Pympler==1.0.1
pynndescent==0.5.6
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
rake-nltk==1.0.6
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.5.4
segtok==1.5.11
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.5
seqeval==1.2.2
shortuuid==1.0.8
simpletransformers==0.63.4
six==1.16.0
smmap==5.0.0
streamlit==1.5.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.49.0
traitlets==4.3.3
transformers==4.15.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
umap-learn==0.5.2
urllib3==1.26.8
validators==0.18.2
wandb==0.12.10
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
xxhash==2.0.2
yake==0.4.8
yarg==0.1.9
yaspin==2.1.0
zipp==3.6.0
